Rename ReduceStore -> Store
